[Analyzers][CPP]Turn on C26451 and fix code (#21230)
* Turn on rule as error and fixing code

* Add packages to external include paths

* Disable warrnings on external code.

+++ src/modules/videoconference/VideoConferenceShared/SerializedSharedMemory.cpp
@@ -83,7 +83,7 @@ std::optional<SerializedSharedMemory> SerializedSharedMemory::create(const std::
     }
 
     // We need an extra byte for locking if it's not readonly
-    const ULARGE_INTEGER UISize{ .QuadPart = size + !read_only };
+    const ULARGE_INTEGER UISize{ .QuadPart = static_cast<uint64_t>(size) + !read_only };
 
     wil::unique_handle hMapFile{ CreateFileMappingW(INVALID_HANDLE_VALUE,
                                                     maybe_attributes ? maybe_attributes : &sa,
